Module Odoc_html.Html_page

module Html = Tyxml.Html

Supported languages for printing code parts.

Page creator

val make : config:Config.t -> url:Odoc_document.Url.Path.t -> header:Html_types.flow5_without_header_footer Html.elt list -> breadcrumbs:Types.breadcrumb list -> toc:Types.toc list -> uses_katex:bool -> Html_types.div_content Html.elt list -> Odoc_document.Renderer.page list -> Odoc_document.Renderer.page

make ?theme_uri (body, children) calls "the page creator" to turn body into an [ `Html ] elt. If theme_uri is provided, it will be used to locate the theme files, otherwise the HTML output directory is used.

val make_src : config:Config.t -> url:Odoc_document.Url.Path.t -> breadcrumbs:Types.breadcrumb list -> header:Html_types.flow5_without_header_footer Html.elt list -> string -> Html_types.div_content Html.elt list -> Odoc_document.Renderer.page

make ?theme_uri (body, children) calls "the page creator" to turn body into an [ `Html ] elt. If theme_uri is provided, it will be used to locate the theme files, otherwise the HTML output directory is used.